+++ middle-server/tencent-cloud-server/src/cloud_manager/mod.rs
@@ -1,0 +1,7 @@
+pub mod common_parameters;
+pub mod cloud_manager;
+mod signature;
+pub mod entities;
+
+
+pub use cloud_manager::TencentCloudManager;
