+++ middle-server/tencent-cloud-server/src/lib.rs
@@ -1,7 +1,6 @@
 pub mod axum_starter;
 pub mod cdn;
-pub mod cloud_manager;
-mod common_parameters;
 pub mod config;
 pub mod error;
 pub mod requester;
+mod cloud_manager;
