+++ src/serves/backend/bakery_mansion/controllers.rs
@@ -2,7 +2,7 @@ use bakery_logic::{impletements::BakeryLogic, view::MansionResp};
 use checker::CheckExtract;
 use persistence::mongodb::MongoDatabaseOperate;
 use resp_result::resp_try;
-use tencent_cloud_server::cloud_manager::TcCloudManager;
+use tencent_cloud_server::cloud_manager::TencentCloudManager;
 use tracing::instrument;
 
 use super::{
@@ -17,7 +17,7 @@ use crate::router::BakeryMansionBackend;
 impl BakeryMansionBackend {
     #[instrument(skip(db, tc_cloud), ret)]
     pub async fn save_mansion(
-        db: MongoDatabaseOperate, tc_cloud: TcCloudManager,
+        db: MongoDatabaseOperate, tc_cloud: TencentCloudManager,
         CheckExtract(mid, ..): OptionMidCheckerPretreatment,
         CheckExtract(json, ..): MansionBodyCheckerPretreatment,
     ) -> MansionRResult<()> {
@@ -52,7 +52,7 @@ impl BakeryMansionBackend {
 
     #[instrument(ret, skip(db, tc_cloud))]
     pub async fn remove_mansion(
-        db: MongoDatabaseOperate, tc_cloud: TcCloudManager,
+        db: MongoDatabaseOperate, tc_cloud: TencentCloudManager,
         CheckExtract(mid, ..): MidCheckerPretreatment,
     ) -> MansionRResult<()> {
         resp_try(async {
