+++ src/serves/backend/ceobe_operation/announcement/controllers.rs
@@ -9,7 +9,7 @@ use persistence::{
     ceobe_operate::models::announcement, mysql::SqlDatabaseOperate,
 };
 use resp_result::resp_try;
-use tencent_cloud_server::cloud_manager::TcCloudManager;
+use tencent_cloud_server::cloud_manager::TencentCloudManager;
 use tracing::instrument;
 
 use super::error::{AnnouncementRespResult, CeobeOperationAnnouncementError};
@@ -39,7 +39,7 @@ impl CeobeOperationAnnouncement {
     #[instrument(ret, skip(db, tc_cloud))]
     // 更新公告列表
     pub async fn update_announcement_list(
-        db: SqlDatabaseOperate, tc_cloud: TcCloudManager,
+        db: SqlDatabaseOperate, tc_cloud: TencentCloudManager,
         CheckExtract(announcements): UpdateAnnouncementCheck,
     ) -> AnnouncementRespResult<()> {
         resp_try(async {
