+++ src/serves/backend/ceobe_operation/resource/controller.rs
@@ -6,7 +6,7 @@ use persistence::{
     ceobe_operate::models::resource, mysql::SqlDatabaseOperate,
 };
 use resp_result::{resp_try, rtry, RespResult};
-use tencent_cloud_server::cloud_manager::TcCloudManager;
+use tencent_cloud_server::cloud_manager::TencentCloudManager;
 use tracing::instrument;
 
 use super::error::{ResourceError, ResourceRResult};
@@ -17,7 +17,7 @@ type ResourceUploadCheck = JsonCheckExtract<resource::Checker, ResourceError>;
 impl CeobeOpResource {
     #[instrument(ret, skip(db, tc_cloud))]
     pub async fn upload_resource(
-        db: SqlDatabaseOperate, tc_cloud: TcCloudManager,
+        db: SqlDatabaseOperate, tc_cloud: TencentCloudManager,
         CheckExtract(resource): ResourceUploadCheck,
     ) -> ResourceRResult<()> {
         resp_try(async {
